Improvements to makesymlinks

+++ makesymlinks.sh
@@ -13,24 +13,23 @@ files="zshrc vimrc vim oh-my-zsh"    # list of files/folders to symlink in homed
 
 ##########
 
+already_existed=0
+if [ -d $dir ]; then
+	already_existed=1
+fi
+
 # create dotfiles_old in homedir
-echo -n "Creating $olddir for backup of any existing dotfiles in ~ ..."
 mkdir -p $olddir
-echo "done"
 
 # change to the dotfiles directory
-echo -n "Changing to the $dir directory ..."
 cd $dir
-echo "done"
 
-git pull
-git submodule update --init
+git pull --quiet
+git submodule update --init --quiet
 
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks from the homedir to any files in the ~/dotfiles directory specified in $files
 for file in $files; do
-    echo "Moving any existing dotfiles from ~ to $olddir"
     mv ~/.$file ~/dotfiles_old/
-    echo "Creating symlink to $file in home directory."
     ln -s $dir/$file ~/.$file
 done
 
@@ -64,8 +63,12 @@ else
 fi
 }
 
-read -p "Should I install oh-my-zsh and change the shell? (y/n) " -n 1;
-echo "";
-if [[ $REPLY =~ ^[Yy]$ ]]; then
-	install_zsh
-fi;
+if [ already_existed = 0 ]; then
+	read -p "Should I install oh-my-zsh and change the shell? (y/n) " -n 1;
+	echo "";
+	if [[ $REPLY =~ ^[Yy]$ ]]; then
+		install_zsh
+	fi
+fi
+
+echo "Done makesymlinks.sh"
